q1_auto_email_responder/requirements.txt
# Intelligent Email Response System Requirements
# Python 3.11 compatible versions

# Core LangChain and LangGraph (latest compatible as of July 2024)
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.2,<0.3
langgraph>=0.2.20,<0.3

# Gmail MCP Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email Processing
email-validator==2.1.0
python-dotenv==1.0.0

# Vector Database and Embeddings
chromadb==0.4.15
sentence-transformers>=2.3.0

# Caching and Performance
redis==4.6.0
cachetools==5.3.2

# Async Support
aiohttp==3.8.6
asyncio-mqtt==0.16.1

# Data Processing
pandas==2.0.3
numpy==1.24.4

# Configuration and Logging
pydantic>=2,<3
pydantic-settings>=2.4.0,<3.0.0
loguru==0.7.2

# Web Framework for API
fastapi==0.103.2
uvicorn==0.23.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional dependencies for Python 3.11 compatibility
typing-extensions>=4.12.0 